Create test_percy_snapshot.py
Create a single app that generate a webpage containing all the screenshots taken in the test_usage.py. We then take a screenshot of everything using percy, and this is used to compare differences.

We employ this method because it is currently impossible for percy to render canvas elements (since it directly crawls the DOM), so this is the only way to proceed